Aborting the redirect
