+++ src/main/java/github/jdrost1818/controller/ExampleRestController.java
@@ -1,6 +1,7 @@
 package github.jdrost1818.controller;
 
 import github.jdrost1818.model.authentication.SessionUser;
+import github.jdrost1818.model.authentication.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -21,6 +22,12 @@ public class ExampleRestController {
     @RequestMapping(value = "/json", method = RequestMethod.GET)
     public String fetchRandomJson() {
         return "Current User: " +  sessionUser.getName();
+    }
+
+    @ResponseStatus(HttpStatus.OK)
+    @RequestMapping(value = "/curuser", method = RequestMethod.GET)
+    public User fetchCurrentUser() {
+        return sessionUser.getCurrentUser();
     }
 
 }
