+++ src/main/java/github/jdrost1818/controller/StaticController.java
@@ -1,8 +1,12 @@
 package github.jdrost1818.controller;
 
+import github.jdrost1818.model.authentication.SessionUser;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+
+import static java.util.Objects.nonNull;
 
 /**
  * Created by JAD0911 on 3/24/2016.
@@ -10,16 +14,19 @@ import org.springframework.web.bind.annotation.RequestMethod;
 @Controller
 public class StaticController {
 
-    private static final String HOME = "login.html";
+    @Autowired
+    protected SessionUser sessionUser;
+
+    private static final String LOGIN = "login.html";
 
     private static final String DASHBOARD = "dashboard.html";
 
     @RequestMapping(value = "/", method = RequestMethod.GET)
     public String home() {
-        return HOME;
+        return LOGIN;
     }
 
-    @RequestMapping(value = "/dashboard", method = RequestMethod.GET)
+    @RequestMapping(value = "/home", method = RequestMethod.GET)
     public String dashboard() {
         return DASHBOARD;
     }
