+++ src/main/java/github/jdrost1818/model/authentication/SessionUser.java
@@ -17,6 +17,10 @@ public class SessionUser implements Serializable {
 
     private User currentUser;
 
+    public User getCurrentUser() {
+        return currentUser;
+    }
+
     public void setCurrentUser(User currentUser) {
         this.currentUser = currentUser;
     }
