+++ src/main/resources/static/js/controllers/header.js
@@ -3,12 +3,12 @@ app.directive('header', () => {
         restrict: 'A',
         replace: true,
         templateUrl: "/js/directives/header.html",
-        controller:  ['$scope', '$http', ($scope, $http) => {
+        controller:  ['$scope', '$http', '$location', ($scope, $http, $location) => {
 
             // Gets the logged in user - if there is one
             $http.get("/user").success(function(data) {
                 $scope.user = data;
-                $scope.authenticated = true;
+                $scope.authenticated = data != "";
             }).error(function() {
                 $scope.user = "N/A";
                 $scope.authenticated = false;
@@ -16,10 +16,10 @@ app.directive('header', () => {
 
             $scope.logout = function() {
                 $http.post('/logout', {}).success(function() {
+                    $scope.user = {};
                     $scope.authenticated = false;
                     $location.path("/");
                 }).error(function(data) {
-                    console.log("Logout failed");
                     $scope.authenticated = false;
                 });
             };
